LPS-96126 Move com.liferay.portal.service.Resource*Test to module and adapt to aquillian test
